pyproject.toml
[tool.poetry]
name = "controle-financeiro"
version = "0.4.0"
description = "Controle finaceiro de contas"
authors = ["Douglas Panhota <douglaspands@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
python-decouple = "^3.4"
python-dateutil = "^2.8.1"
django-crispy-forms = "^1.11.2"
prompt-toolkit = "^3.0.18"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
django-debug-toolbar = "^3.2"
mypy = "^0.812"
ipython = "^7.22.0"
black = {version = "^20.8b1", allow-prereleases = true}
coverage = "^5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
runserver = "scripts:runserver"
test = "scripts:test"
coverage = "scripts:coverage"
testcov = "scripts:testcov"
lint = "scripts:lint"
migrate = "scripts:migrate"
startapp = "scripts:startapp"
makemigrations = "scripts:makemigrations"
requirements = "scripts:requirements"
createsuperuser = "scripts:createsuperuser"
collectstatic = "scripts:collectstatic"
shell = "scripts:djangoshell"
dbshell = "scripts:dbshell"
command = "scripts:command"
migrate_remove = "scripts:migrate_remove"
